+++ Scripts/Remote/Show-WinRMConfig.ps1
@@ -101,6 +101,7 @@ param (
 
 $ErrorActionPreference = "Stop"
 $InformationPreference = "Continue"
+. $PSScriptRoot\WinRMSettings.ps1
 
 $WinRM = Get-Service -Name WinRM
 Write-Information -Tags "User" -MessageData "INFO: WinRM service status=$($WinRM.Status) startup=$($WinRM.StartType)"
@@ -146,18 +147,20 @@ if ($WinRM.Status -ne "Running")
 # MSDN: Select-Object, beginning in PowerShell 6,
 # it is no longer required to include the Property parameter for ExcludeProperty to work.
 
-# winrm get winrm/config
-Write-Information -Tags "Project" -MessageData "INFO: Showing all enabled session configurations (short version)"
-Get-PSSessionConfiguration | Where-Object -Property Enabled -EQ True |
-Select-Object -Property Name, Enabled, PSVersion, Architecture, SupportsOptions, lang, AutoRestart, RunAsUser, RunAsPassword, Permission
-
-# winrm enumerate winrm/config/listener
-Write-Information -Tags "Project" -MessageData "INFO: Showing configured listeners"
-Get-WSManInstance -ResourceURI winrm/config/Listener -Enumerate |
-Select-Object -ExcludeProperty cfg, xsi
-
 if ($Server)
 {
+	# winrm get winrm/config
+	Write-Information -Tags "Project" -MessageData "INFO: Showing all enabled session configurations (short version)"
+	Get-PSSessionConfiguration | Where-Object -Property Enabled -EQ True |
+	Select-Object -Property Name, Enabled, PSVersion, SDKVersion, Architecture,
+	lang, Capability, SupportsOptions, AutoRestart, RunAsUser, RunAsPassword,
+	RunAsVirtualAccount, RunAsVirtualAccountGroups, Permission
+
+	# winrm enumerate winrm/config/listener
+	Write-Information -Tags "Project" -MessageData "INFO: Showing configured listeners"
+	Get-WSManInstance -ResourceURI winrm/config/Listener -Enumerate |
+	Select-Object -ExcludeProperty cfg, xsi
+
 	# winrm get winrm/config/service
 	Write-Information -Tags "Project" -MessageData "INFO: Showing server configuration"
 	Get-WSManInstance -ResourceURI winrm/config/Service |