+++ Readme/CHANGELOG.md
@@ -54,7 +54,7 @@ If you would like to see fresh changes done since last release you can do so on 
   - `Enable-WinRMServer.ps1` script to configure WinRM server for remoting
   - `Disable-WinRMServer.ps1` script to disable WinRM remoting
   - `Set-WinRMClient.ps1` script to configure client for remoting
-  - `Install-SslCertificate.ps1` script to install, create and export SSL certificate for remoting
+  - `Register-SslCertificate.ps1` script to install, create and export SSL certificate for remoting
   - `Connect-Computer.ps1` script to initialize connection to remote computer
 
 - Modules